v.pradinė 3 - galutinio pažymio apskaičiavimas

+++ main.h
@@ -9,6 +9,7 @@
 #include <ios>          // streamsize
 #include <algorithm>    // all_of
 #include <cctype>       // isalpha
+#include <numeric>      // accumulate
 
 struct studentas 
 {
@@ -16,7 +17,7 @@ struct studentas
     std::string pavarde {};
     std::vector<int> namuDarbai;
     int egzaminas {};
-    int galutinis {};
+    double galutinis {};
 };
 
 struct studentasNode {
